src: split code into modules
