+++ src/Err.hs
@@ -1,0 +1,8 @@
+module Err 
+    ( Err(..)
+    ) where
+
+data Err = Err { reason :: String }
+
+instance Show Err where
+  show e = "Error: " ++ (reason e)
