+++ src/Tokenize.hs
@@ -1,0 +1,21 @@
+module Tokenize
+    ( tokenize
+    ) where
+
+-- adapted from: http://bluebones.net/2007/01/replace-in-haskell/
+replace :: Eq a => [a] -> [a] -> [a] -> [a]
+replace _ _ [] = []
+replace find repl s =
+    if take (length find) s == find
+        then repl ++ (replace find repl (drop (length find) s))
+        else [head s] ++ (replace find repl (tail s))
+
+-- pipeline operator
+x |> f = f x
+
+tokenize :: String -> [String]
+tokenize expr = expr
+  |> replace "(" " ( "
+  |> replace ")" " ) "
+  |> words
+
